+++ Sisapco/src/main/java/co/com/sisapco/repository/CotizacionesRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import co.com.sisapco.entity.Cotizaciones;
 
 @Repository
-public interface CotizacionesRepository extends CrudRepository<Cotizaciones, Long>{
+public interface CotizacionesRepository extends CrudRepository<Cotizaciones, Integer>{
 	
 	@Query("SELECT u FROM Cotizaciones u WHERE u.copNit = ?1 ORDER BY u.cotId DESC")
 	public Iterable<Cotizaciones> findByCotizacionesNit(int copNit);
