+++ Sisapco/src/main/java/co/com/sisapco/repository/PerfilRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import co.com.sisapco.entity.Perfil;
 
 @Repository
-public interface PerfilRepository extends CrudRepository<Perfil, Long> {
+public interface PerfilRepository extends CrudRepository<Perfil, Integer> {
 	
 	 @Query("SELECT u FROM Perfil u WHERE u.perId = ?1 AND u.perEstado = 1")	
 	 public Optional<Perfil> findByPerfilId(int perfilId);
